chore: separate highlight fns
