+++ src/lib/textCSS.ts
@@ -1,3 +1,4 @@
+import { highlightCSS } from "./highlightText";
 import { roundToTwoDecimals } from "./utils";
 
 const HEIGHT_TO_WIDTH_RATIO = 1.1547005;
@@ -254,16 +255,6 @@ export interface CSSInputValues {
   mediaQuery_2: number;
   mediaQuery_3: number;
   textColor: string;
-}
-
-export function highlightCSS(cssString: string) {
-  return cssString
-    .replace(/([.#]?\w[\w-]*)\s*{/g, '<span class="css-selector">$1</span> {')
-    .replace(
-      /([a-z-]+)(\s*:\s*)([^;]+)(;?)/g,
-      '<span class="css-property">$1</span>$2<span class="css-value">$3</span>$4',
-    )
-    .replace(/({|})/g, '<span class="css-brace">$1</span>');
 }
 
 export function generateCSSText({
