+++ src/lib/textHTML.ts
@@ -1,3 +1,5 @@
+import { highlightHTML } from "./highlightText";
+
 export function generateHexagonHTML(hexagonNumber: number) {
   const hexagon = `
     <div class="hexagon__outer">
@@ -29,25 +31,6 @@ export function wrapHTML(html: string) {
 </div>
 `;
   return wrapper;
-}
-
-export function replaceSymbolsInHTMLForDisplay(htmlString: string) {
-  return htmlString.replace(/[<>]/g, (match) =>
-    match === "<" ? "&lt;" : "&gt;",
-  );
-}
-
-export function highlightHTML(htmlString: string) {
-  return replaceSymbolsInHTMLForDisplay(htmlString).replace(
-    /(&lt;\/?)([\w-]+)([\s\S]*?)(&gt;)/g,
-    (_match, p1, p2, p3, p4) => {
-      const attrHighlight = p3.replace(
-        /(\b(class|id))(=")([\w-]+)(")/g,
-        '<span class="html-attribute">$2</span><span class="html-symbol">$3</span><span class="html-attribute">$4</span><span class="html-symbol">$5</span>',
-      );
-      return `<span class="html-symbol">${p1}</span><span class="html-element">${p2}</span>${attrHighlight}<span class="html-symbol">${p4}</span>`;
-    },
-  );
 }
 
 export interface HTMLInputValues {
